+++ pyabc/epsilon/epsilon.py
@@ -1,7 +1,7 @@
 import scipy as sp
 import pandas as pd
 import logging
-from typing import Callable, Dict, List, Union
+from typing import Callable, List, Union
 from .base import Epsilon
 from ..weighted_statistics import weighted_quantile
 
@@ -138,7 +138,7 @@ class QuantileEpsilon(Epsilon):
     def initialize(self,
                    t: int,
                    get_weighted_distances: Callable[[], pd.DataFrame],
-                   get_all_records: Callable[[], List[Dict]],
+                   get_all_records: Callable[[], List[dict]],
                    max_nr_populations: int,
                    acceptor_config: dict):
         if self._initial_epsilon != 'from_sample':
@@ -183,7 +183,7 @@ class QuantileEpsilon(Epsilon):
     def update(self,
                t: int,
                get_weighted_distances: Callable[[], pd.DataFrame],
-               get_all_records: Callable[[], List[Dict]],
+               get_all_records: Callable[[], List[dict]],
                acceptance_rate: float,
                acceptor_config: dict):
         """